+++ src/main/java/roomescape/reservation/application/service/WaitingReservationCommandService.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.application.service;
+
+import roomescape.reservation.application.dto.CreateReservationServiceRequest;
+import roomescape.reservation.domain.WaitingReservation;
+
+public interface WaitingReservationCommandService {
+
+    WaitingReservation create(CreateReservationServiceRequest request);
+
+    void delete(Long id);
+}
