+++ src/test/java/roomescape/reservation/application/service/ReservationCommandServiceImplTest.java
@@ -14,7 +14,6 @@ import roomescape.reservation.application.dto.CreateReservationServiceRequest;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationRepository;
-import roomescape.reservation.domain.WaitingReservation;
 import roomescape.reservation.domain.WaitingReservationRepository;
 import roomescape.reservation.exception.PastDateReservationException;
 import roomescape.reservation.exception.PastTimeReservationException;
@@ -190,159 +189,6 @@ class ReservationCommandServiceImplTest {
         assertThatThrownBy(() -> reservationCommandService.delete(id))
                 .isInstanceOf(NotFoundException.class)
                 .hasMessage("[RESERVATION] not found. params={Long=-1}");
-    }
-
-    @Test
-    @DisplayName("예약이 존재할 때, 예약 대기를 추가한다")
-    void createWaitingReservation() {
-        // given
-        final ReservationTime reservationTime = createAndSaveReservationTime(LocalTime.of(10, 0));
-        final Theme theme = createAndSaveTheme();
-        final User user = createAndSaveUser();
-
-        final CreateReservationServiceRequest requestDto = createReservationRequest(
-                user.getId(),
-                LocalDate.of(2025, 8, 5),
-                reservationTime.getId(),
-                theme.getId());
-
-        final Reservation reservation = reservationRepository.save(
-                Reservation.withoutId(
-                        user.getId(),
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme));
-
-        // when
-        final WaitingReservation waitingReservation = reservationCommandService.createWaitingReservation(requestDto);
-
-        // then
-        final WaitingReservation found = waitingReservationRepository.findById(waitingReservation.getId())
-                .orElseThrow(NoSuchElementException::new);
-
-        assertThat(waitingReservation).isEqualTo(found);
-        assertThat(waitingReservation.getId()).isEqualTo(found.getId());
-        assertThat(waitingReservation.getUserId()).isEqualTo(found.getUserId());
-        assertThat(waitingReservation.getDate()).isEqualTo(found.getDate());
-        assertThat(waitingReservation.getTime()).isEqualTo(found.getTime());
-        assertThat(waitingReservation.getTheme()).isEqualTo(found.getTheme());
-    }
-
-    @Test
-    @DisplayName("중복된 예약 대기가 존재할 때, 예외가 발생한다")
-    void createWaitingReservationDuplication() {
-        // given
-        final ReservationTime reservationTime = createAndSaveReservationTime(LocalTime.of(10, 0));
-        final Theme theme = createAndSaveTheme();
-        final User user = createAndSaveUser();
-
-        final CreateReservationServiceRequest requestDto = createReservationRequest(
-                user.getId(),
-                LocalDate.of(2025, 8, 5),
-                reservationTime.getId(),
-                theme.getId());
-
-        final Reservation reservation = reservationRepository.save(
-                Reservation.withoutId(
-                        user.getId(),
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme));
-
-        // when
-        final WaitingReservation waitingReservation = reservationCommandService.createWaitingReservation(requestDto);
-        assertThatThrownBy(() -> reservationCommandService.createWaitingReservation(requestDto))
-                .isInstanceOf(DuplicateException.class)
-                .hasMessageContainingAll(
-                        "RESERVATION_WAITING already exists.",
-                        "params={ReservationDate=ReservationDate(value=",
-                        "DomainTerm=THEME_ID, DomainTerm=RESERVATION_TIME_ID}"
-                );
-
-    }
-
-    @Test
-    @DisplayName("예약 대기를 삭제할 수 있다")
-    void deleteWaitingReservation() {
-        // given
-        final ReservationTime reservationTime = createAndSaveReservationTime(LocalTime.of(10, 0));
-        final Theme theme = createAndSaveTheme();
-        final User user = createAndSaveUser();
-
-        final WaitingReservation waitingReservation = waitingReservationRepository.save(
-                WaitingReservation.withoutId(
-                        user.getId(),
-                        1,
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme
-                ));
-
-        // when
-        Long id = waitingReservation.getId();
-        reservationCommandService.deleteWaiting(id);
-
-        // then
-        assertThat(waitingReservationRepository.findById(id)).isEmpty();
-    }
-
-    @Test
-    @DisplayName("예약 대기 삭제 시 뒤의 순서들이 자동으로 업데이트된다")
-    void deleteWaitingReservationUpdatesOrder() {
-        //given
-        final ReservationTime reservationTime = createAndSaveReservationTime(LocalTime.of(10, 0));
-        final Theme theme = createAndSaveTheme();
-        final User user = createAndSaveUser();
-
-        final WaitingReservation waiting1 = waitingReservationRepository.save(
-                WaitingReservation.withoutId(
-                        user.getId(),
-                        1,
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme));
-
-        final WaitingReservation waiting2 = waitingReservationRepository.save(
-                WaitingReservation.withoutId(
-                        user.getId(),
-                        2,
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme));
-
-        final WaitingReservation waiting3 = waitingReservationRepository.save(
-                WaitingReservation.withoutId(
-                        user.getId(),
-                        3,
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme));
-
-        final WaitingReservation waiting4 = waitingReservationRepository.save(
-                WaitingReservation.withoutId(
-                        user.getId(),
-                        4,
-                        ReservationDate.from(LocalDate.of(2025, 8, 5)),
-                        reservationTime,
-                        theme));
-
-        //when
-        reservationCommandService.deleteWaiting(waiting2.getId());
-
-        //then
-        assertThat(waitingReservationRepository.findById(waiting2.getId())).isEmpty();
-
-        final WaitingReservation updatedWaiting1 = waitingReservationRepository.findById(waiting1.getId())
-                .orElseThrow(NoSuchElementException::new);
-        assertThat(updatedWaiting1.getWaitingOrder()).isEqualTo(1);
-
-        final WaitingReservation updatedWaiting3 = waitingReservationRepository.findById(waiting3.getId())
-                .orElseThrow(NoSuchElementException::new);
-        assertThat(updatedWaiting3.getWaitingOrder()).isEqualTo(2);
-
-        final WaitingReservation updatedWaiting4 = waitingReservationRepository.findById(waiting4.getId())
-                .orElseThrow(NoSuchElementException::new);
-        assertThat(updatedWaiting4.getWaitingOrder()).isEqualTo(3);
     }
 
     // Helper 메서드들
